refactor: todos 분리
- map 구현 상세 감춤
- 폴더 캡슐화 수정
